Upload
Cleaned up unused vars
Added ability to see files that have already uploaded when you try to update your submit
